+++ src/dipdup/config/__init__.py
@@ -90,7 +90,7 @@ type WsUrl = Annotated[str, BeforeValidator(lambda v: _valid_url(v, ws=True))]
 _logger = logging.getLogger(__name__)
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class SqliteDatabaseConfig:
     """
     SQLite connection config
@@ -124,7 +124,7 @@ class SqliteDatabaseConfig:
         return 1
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class PostgresDatabaseConfig:
     """Postgres database connection config
 
@@ -176,7 +176,7 @@ class PostgresDatabaseConfig:
         }
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class HttpConfig:
     """Advanced configuration of HTTP client
 
@@ -212,7 +212,7 @@ class HttpConfig:
     alias: str | None = None
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class ResolvedHttpConfig:
     __doc__ = HttpConfig.__doc__
 
@@ -291,7 +291,7 @@ class DatasourceConfig(ABC, NameMixin):
     http: HttpConfig | None = None
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class HandlerConfig(CallbackMixin, ParentMixin['IndexConfig']):
     """Base class for index handlers
 
@@ -303,7 +303,7 @@ class HandlerConfig(CallbackMixin, ParentMixin['IndexConfig']):
         ParentMixin.__post_init__(self)
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class IndexTemplateConfig(NameMixin):
     """Index template config
 
@@ -322,7 +322,7 @@ class IndexTemplateConfig(NameMixin):
     last_level: int = 0
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class IndexConfig(ABC, NameMixin, ParentMixin['ResolvedIndexConfigU']):
     """Index config
 
@@ -363,7 +363,7 @@ class IndexConfig(ABC, NameMixin, ParentMixin['ResolvedIndexConfigU']):
             datasource.pop('buffer_size', None)
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class HasuraConfig:
     """Config for the Hasura integration.
 
@@ -402,7 +402,7 @@ class HasuraConfig:
         return {}
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class JobConfig(NameMixin):
     """Job schedule config
 
@@ -429,7 +429,7 @@ class JobConfig(NameMixin):
         NameMixin.__post_init__(self)
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class SentryConfig:
     """Config for Sentry integration.
 
@@ -449,7 +449,7 @@ class SentryConfig:
     debug: bool = False
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class PrometheusConfig:
     """Config for Prometheus integration.
 
@@ -463,7 +463,7 @@ class PrometheusConfig:
     update_interval: float = 1.0
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class HookConfig(CallbackMixin):
     """Hook config
 
@@ -488,7 +488,7 @@ class HookConfig(CallbackMixin):
                 yield package, obj
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class SystemHookConfig(HookConfig):
     __doc__ = HookConfig.__doc__
 
@@ -522,7 +522,7 @@ SYSTEM_HOOKS = {
 }
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class ApiConfig:
     """Management API config
 
@@ -535,7 +535,7 @@ class ApiConfig:
 
 
 # NOTE: Should be the only place where extras are allowed
-@dataclass(config=ConfigDict(extra='allow'), kw_only=True)
+@dataclass(config=ConfigDict(extra='allow', defer_build=True), kw_only=True)
 class AdvancedConfig:
     """This section allows users to tune some system-wide options, either experimental or unsuitable for generic configurations.
 
@@ -559,7 +559,7 @@ class AdvancedConfig:
     alt_operation_matcher: bool = False
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class DipDupConfig:
     """DipDup project configuration file
 
@@ -634,6 +634,9 @@ class DipDupConfig:
         )
 
         try:
+            # from pydantic.dataclasses import rebuild_dataclass
+            # rebuild_dataclass(cls, force=True)
+
             config = TypeAdapter(cls).validate_python(config_json)
         except ConfigurationError:
             raise
@@ -1189,15 +1192,6 @@ ResolvedIndexConfigU = TezosIndexConfigU | EvmIndexConfigU | StarknetIndexConfig
 IndexConfigU = ResolvedIndexConfigU | IndexTemplateConfig
 
 
-def _reload_dataclass(cls: type[Any]) -> type[Any]:
-    """Reload dataclass to apply new annotations"""
-    try:
-        return dataclass(cls, config=cls.__pydantic_config__, kw_only=True)
-    # NOTE: The first attempt fails with "dictionary changed size" due to how deeply fucked up this hack is.
-    except RuntimeError:
-        return dataclass(cls, config=cls.__pydantic_config__, kw_only=True)
-
-
 def _patch_annotations() -> None:
     """Patch dataclass annotations in runtime to allow using aliases in config files.
 
@@ -1213,7 +1207,7 @@ def _patch_annotations() -> None:
         (self.__name__, self),
     )
 
-    for name, submodule in submodules:
+    for _, submodule in submodules:
         if not submodule.__name__.startswith('dipdup.config'):
             continue
 
@@ -1235,17 +1229,10 @@ def _patch_annotations() -> None:
                     body, after = body.split(']', 1)
                     unwrapped = f'{before}str | {body}{after}'
 
-                if annotation != unwrapped:
+                if unwrapped != annotation:
                     value.__annotations__[name] = unwrapped
-
-            setattr(
-                submodule,
-                attr,
-                _reload_dataclass(value),
-            )
-
-    # NOTE: Finally, reload the root config itself.
-    self.DipDupConfig = _reload_dataclass(DipDupConfig)  # type: ignore[attr-defined]
+                    value.__pydantic_fields__[name].annotation = unwrapped  # type: ignore[assignment]
+                    value.__dataclass_fields__[name].type = unwrapped
 
 
 _patch_annotations()