+++ src/dipdup/config/evm_node.py
@@ -11,7 +11,7 @@ from dipdup.config import Url
 from dipdup.config import WsUrl
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class EvmNodeDatasourceConfig(DatasourceConfig):
     """EVM node datasource config