+++ src/dipdup/config/substrate_events.py
@@ -22,7 +22,7 @@ if TYPE_CHECKING:
     from dipdup.subscriptions import Subscription
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class SubstrateEventsHandlerConfig(HandlerConfig):
     """Subsquid event handler
 
@@ -49,7 +49,7 @@ class SubstrateEventsHandlerConfig(HandlerConfig):
         yield 'event', f'SubstrateEvent[{event_cls}]'
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class SubstrateEventsIndexConfig(SubstrateIndexConfig):
     """Subsquid datasource config