+++ src/dipdup/config/substrate_subscan.py
@@ -9,7 +9,7 @@ from dipdup.config import DatasourceConfig
 from dipdup.config import HttpConfig
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class SubstrateSubscanDatasourceConfig(DatasourceConfig):
     """Subscan datasource config