+++ src/dipdup/config/substrate_subsquid.py
@@ -10,7 +10,7 @@ from dipdup.config import HttpConfig
 from dipdup.config import Url
 
 
-@dataclass(config=ConfigDict(extra='forbid'), kw_only=True)
+@dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
 class SubstrateSubsquidDatasourceConfig(DatasourceConfig):
     """Subsquid datasource config